Add check for contents file ``pre text''
This pre text is described in passing in the contents indices format
document as *possibly* existing, and that the first row of the table
proper *should* have the headers ``FILE'' and ``LOCATION''. I am going
on the assumption that *if* the file has pre text, then it *must* have
the table headers as well.